fix: clear the last seven Codacy findings
The stream and wire modules crossed the file length limit, almost
entirely from their test blocks, so both become directory modules with
the tests in a sibling tests.rs. No code changed, only where it lives.

The ResizeObserver test stub used a class through vi.stubGlobal, which
the analyzers reject in a setup file; it is a plain function stub
assigned on globalThis now.

The workspace hoist left src-tauri/Cargo.lock behind, superseded by the
root workspace lockfile cargo actually maintains; the stale copy is
deleted. Cargo.lock is excluded from Codacy analysis because it is a
generated file and Rust dependency advisories are already enforced by
the cargo-deny gate in CI with a reviewed policy, so the duplicate
scanner only adds findings nobody can act on from a lockfile.
